requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.1
keybert
sentence-transformers
pandas>=2.0.0
langchain
langchain-groq 